03.dj_user_demo/backend/apps/users/views.py
from django.shortcuts import render
from rest_framework import viewsets
from rest_framework.response import Response
from rest_framework import status

from apps.users.serializers import RegisterSerializer


# Create your views here.

class RegisterView(viewsets.ViewSet):
    def post(self, request):
        user = request.data
        serializer = self.serializer_class(data=user)
        serializer.is_valid(raise_exception=True)
        serializer.save()

        user_data = serializer.data
        return Response(user_data, status=status.HTTP_201_CREATED)
